refactor(core): address #414 review — clone-mutate, shared Debug/env bool, split tests
- with_token_provider: clone ConfigInner then replace token_provider
- ConfigBuilder Debug reuses ConfigInner redacted output
- unify OPENLARK_ENABLE_* bool parsing via parse_env_bool
- move config unit tests to config_tests.rs (prod file under 1k)
